Clean testing asserters
