+++ pandas/_testing/__init__.py
@@ -58,7 +58,6 @@ from pandas._testing.asserters import (
     assert_indexing_slices_equivalent,
     assert_interval_array_equal,
     assert_is_sorted,
-    assert_is_valid_plot_return_object,
     assert_metadata_equivalent,
     assert_numpy_array_equal,
     assert_period_array_equal,
@@ -565,7 +564,6 @@ __all__ = [
     "assert_indexing_slices_equivalent",
     "assert_interval_array_equal",
     "assert_is_sorted",
-    "assert_is_valid_plot_return_object",
     "assert_metadata_equivalent",
     "assert_numpy_array_equal",
     "assert_period_array_equal",
